에러 핸들링 설정

+++ sehee/controllers/postsController.js
@@ -2,38 +2,74 @@ const postsService = require("../services/postsService");
 
 // 게시글 등록
 const createPost = async (req, res) => {
-  const { title, content, post_image, user_id } = req.body;
-
-  await postsService.createPost(title, content, post_image, user_id);
-  return res.status(201).json({ message: "postCreated" });
+  try {
+    const { title, content, post_image, user_id } = req.body;
+    if (!title || !content || !post_image || !user_id) {
+      return res.status(400).json({ message: "KEY_ERROR" });
+    }
+    await postsService.createPost(title, content, post_image, user_id);
+    return res.status(201).json({ message: "postCreated" });
+  } catch (err) {
+    console.log(err);
+    return res.status(err.statusCode || 500).json({ message: err.message });
+  }
 };
 
 // 전체 게시글 조회
 const getPosts = async (req, res) => {
-  const posts = await postsService.getPosts();
-  res.status(200).json({ data: posts });
+  try {
+    const posts = await postsService.getPosts();
+    res.status(200).json({ data: posts });
+  } catch (err) {
+    console.log(err);
+    return res.status(err.statusCode || 500).json({ message: err.message });
+  }
 };
 
 // 유저 게시글 조회
 const getPostByUserId = async (req, res) => {
-  const { userId } = req.params;
-  const postUserId = await postsService.getPostByUserId(userId);
-  return res.status(200).json({ data: postUserId });
+  try {
+    const { userId } = req.params;
+    if (!userId) {
+      return res.status(400).json({ message: "KEY_ERROR" });
+    }
+    const postUserId = await postsService.getPostByUserId(userId);
+    return res.status(200).json({ data: postUserId });
+  } catch (err) {
+    console.log(err);
+    return res.status(err.statusCode || 500).json({ message: err.message });
+  }
 };
 
 // 게시글 수정
 const updatePost = async (req, res) => {
-  const { postId } = req.params;
-  const { content } = req.body;
-  await postsService.updatePost(postId, content);
-  return res.status(200).json({ data: "updatePost" });
+  try {
+    const { postId } = req.params;
+    const { content } = req.body;
+    if (!postId || !content) {
+      return res.status(400).json({ message: "INVALID_DATA_INPUT" });
+    }
+    await postsService.updatePost(postId, content);
+    return res.status(200).json({ data: "updatePost" });
+  } catch (err) {
+    console.log(err);
+    return res.status(err.statusCode || 500).json({ message: err.message });
+  }
 };
 
 // 게시글 삭제
 const deletePost = async (req, res) => {
-  const { postId } = req.params;
-  await postsService.deletePost(postId);
-  return res.status(200).json({ message: "postingDeleted" });
+  try {
+    const { postId } = req.params;
+    if (!postId) {
+      return res.status(400).json({ message: "KEY_ERROR" });
+    }
+    await postsService.deletePost(postId);
+    return res.status(200).json({ message: "postingDeleted" });
+  } catch (err) {
+    console.log(err);
+    return res.status(err.statusCode || 500).json({ message: err.message });
+  }
 };
 
 module.exports = {
